requirements.txt
openai>=1.0.0
python-dotenv>=0.19.0
pandas>=1.3.0
numpy>=1.21.0
scipy>=1.7.0
scikit-learn>=0.24.0
sentence-transformers>=2.0.0
matplotlib>=3.4.0
plotly>=5.3.0
pytest>=6.2.0
jsonschema>=4.0.0
pyyaml>=5.4.0 